pyproject.toml
[project]
name = "doc81"
version = "0.1.0"
description = "A tool to help you write better documentation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-frontmatter>=1.1.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Documentation",
]

[project.scripts]
doc81-mcp = "doc81.mcp.main:main"
doc81-mcp-cli = "doc81.mcp.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
mcp = [
    "fastmcp>=2.9.2",
    "typer>=0.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/doc81"]
